any2pdf/requirements.txt
# any2pdf 의존성
python-docx>=1.1.0
markdown>=3.7
beautifulsoup4>=4.12.0
pyhwp>=0.1b15
six>=1.16.0
